Add Jenkins pipeline support.
This commit adds the support for Jenkins Kubernetes Pipeline.
This adds the Jenkinsfile, that has configurations specific to
ICP to make a CI/CD workflow for this application.

Some values are hardcoded for with local URLs, that should be
replaced while running on a ICP cluster that as a different cluster
name.